Moved 404 code out of src
