+++ pages/404.tsx
@@ -1,5 +1,39 @@
-import NotFound from '../src/pages/NotFound' // or wherever your NotFound component is
+import React from 'react'
+
+import Image from 'next/image'
+import { useRouter } from 'next/router'
+
+import GradientButton from '../src/shared/components/Buttons/GradientButton'
+import { DantelionImageUrl } from '../src/shared/utils/imageUrls'
+
+import styles from './404.module.scss'
 
 export default function Custom404() {
-  return <NotFound />
+  const router = useRouter()
+
+  const goBack = () => {
+    if (window.history.length > 1) {
+      router.back()
+    } else {
+      router.push('/')
+    }
+  }
+
+  return (
+    <div className={styles['container']}>
+      <h1 className={styles['title']}>404 - Whoops!</h1>
+      <Image
+        src={DantelionImageUrl}
+        alt="Dantelion"
+        className={styles['image']}
+        width={150}
+        height={150}
+      />
+      <p>You have trespassed into the abyss from which no soul returns!</p>
+
+      <GradientButton bold className={styles['back-button']} onClick={goBack}>
+        Go back
+      </GradientButton>
+    </div>
+  )
 }
